+++ docs/changelog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+### 2025-11-11
+
+- Add missing outputs to the Terraform product module.
+
 ### 2025-09-29
 
 - Terraform product module supports ingress, certificates, oauth2-proxy.